[Add] Location retrieval for weather

+++ app/src/main/java/com/example/photoweather/data/remote/models/Weather.kt
@@ -30,7 +30,7 @@ data class Weather(
     data class Temperature(
         @SerializedName("temp")
         @Expose
-        val temp: Int
+        val temp: Float
     )
 
     data class CountryCode(
